move Halo5 StringId to Gen5 common and add string methods to IMetadataHeader
