+++ Reclaimer.Blam/Blam/Common/Gen5/StringIdGen5.cs
@@ -1,18 +1,18 @@
 ﻿using Reclaimer.IO;
 
-namespace Reclaimer.Blam.Halo5
+namespace Reclaimer.Blam.Common.Gen5
 {
     [FixedSize(8)]
-    public class StringId
+    public class StringIdGen5
     {
-        private readonly MetadataHeader meta;
+        private readonly IMetadataHeader meta;
 
         [Offset(4)]
         public int StringOffset { get; set; }
 
         public string Value => meta.GetStringByOffset(StringOffset);
 
-        public StringId(MetadataHeader meta)
+        public StringIdGen5(IMetadataHeader meta)
         {
             this.meta = meta ?? throw new ArgumentNullException(nameof(meta));
         }
@@ -21,21 +21,21 @@ namespace Reclaimer.Blam.Halo5
     }
 
     [FixedSize(4)]
-    public readonly struct StringHash
+    public readonly struct StringHashGen5
     {
-        private readonly MetadataHeader header;
+        private readonly IMetadataHeader header;
 
         public uint Hash { get; }
 
         public string Value => header.GetStringByHash(Hash);
 
-        public StringHash(EndianReader reader, MetadataHeader header)
+        public StringHashGen5(EndianReader reader, IMetadataHeader header)
         {
             this.header = header;
             Hash = reader.ReadUInt32();
         }
 
-        public static implicit operator string(StringHash value) => value.Value;
+        public static implicit operator string(StringHashGen5 value) => value.Value;
 
         public override string ToString() => Value ?? "[invalid string]";
     }
